refactor: Round pause screen gauge values

+++ core/src/com/github/rjbx/energen/app/LevelScreen.java
@@ -269,7 +269,7 @@ final class LevelScreen extends ScreenAdapter {
         cursor.resetPosition();
         String[] optionStrings = {"RESUME", "EXIT", "OPTIONS"};
         menu.setOptionStrings(Arrays.asList(optionStrings));
-        menu.setPromptString(Align.left, "GAUGE\n" + Constants.HUD_FUEL_LABEL + avatar.getAmmo() + "\n" + Constants.HUD_HEALTH_LABEL + avatar.getHealth() + "\n" + Constants.HUD_STAMINA_LABEL + avatar.getTurbo() + "\n" + Constants.HUD_ENERGY_LABEL + avatar.getEnergy().toString());
+        menu.setPromptString(Align.left, "GAUGE\n" + Constants.HUD_FUEL_LABEL + (int) avatar.getAmmo() + "\n" + Constants.HUD_HEALTH_LABEL + (int) avatar.getHealth() + "\n" + Constants.HUD_STAMINA_LABEL + (int) avatar.getTurbo() + "\n" + Constants.HUD_ENERGY_LABEL + avatar.getEnergy().toString());
         menu.setPromptString(Align.right, "STATS\n" + "Game Time: " + Helpers.millisToString((TimeUtils.nanosToMillis(SaveData.getTotalTime()) + levelUpdater.getUnsavedTime())) + "\nLevel Time: " + Helpers.millisToString(TimeUtils.nanosToMillis(levelUpdater.getTime())) + "\nGame Score: " + (SaveData.getTotalScore() + levelUpdater.getUnsavedScore()) + "\nLevel Score: " + levelUpdater.getScore());
 //        menu.setPromptString(Align.right, "ENERGY\n" + (avatar.getEnergy().name().toLowerCase() + "\n" + SaveData.getEnergies().replace(avatar.getEnergy().name(), "").replace(", ", "\n")).replace("\n\n", "\n").toLowerCase());
         menu.TextAlignment(Align.center);
